Move cloud methods to pkg/secrets
